P4-2538 Add Notification event class
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2538

Add generic event notification class, which all notification events
inherit from, also add shared example test to test notification classification
is set.
