+++ app/models/generic_event/move_notify_premises_of_arrival_in_30_mins.rb
@@ -1,9 +1,5 @@
 class GenericEvent
-  class MoveNotifyPremisesOfArrivalIn30Mins < GenericEvent
+  class MoveNotifyPremisesOfArrivalIn30Mins < Notification
     eventable_types 'Move'
-
-    def event_classification
-      :notification
-    end
   end
 end
