+++ app/models/generic_event/move_notify_premises_of_eta.rb
@@ -1,14 +1,10 @@
 class GenericEvent
-  class MoveNotifyPremisesOfEta < GenericEvent
+  class MoveNotifyPremisesOfEta < Notification
     details_attributes :expected_at
     eventable_types 'Move'
 
     validates :expected_at, presence: true
 
     validates :expected_at, iso_date_time: true
-
-    def event_classification
-      :notification
-    end
   end
 end
