+++ app/models/generic_event/notification.rb
@@ -1,0 +1,7 @@
+class GenericEvent
+  class Notification < GenericEvent
+    def event_classification
+      :notification
+    end
+  end
+end
